Split trace function into separate file
Don't share a file with raise_or_lower_index.
