+++ src/DataStructures/Tensor/EagerMath/RaiseOrLowerIndex.cpp
@@ -55,49 +55,6 @@ void raise_or_lower_index(
   }
 }
 
-template <typename DataType, typename Index0, typename Index1>
-void trace_last_indices(
-    const gsl::not_null<Tensor<DataType, Symmetry<1>, index_list<Index0>>*>
-        trace_of_tensor,
-    const Tensor<DataType, Symmetry<2, 1, 1>,
-                 index_list<Index0, Index1, Index1>>& tensor,
-    const Tensor<DataType, Symmetry<1, 1>,
-                 index_list<change_index_up_lo<Index1>,
-                            change_index_up_lo<Index1>>>& metric) {
-  constexpr auto dimension_of_trace = Index0::dim;
-  constexpr auto dimension_of_metric = Index1::dim;
-  for (size_t i = 0; i < dimension_of_trace; ++i) {
-    trace_of_tensor->get(i) = tensor.get(i, 0, 0) * get<0, 0>(metric);
-    for (size_t j = 0; j < dimension_of_metric; ++j) {
-      if (j != 0) {
-        trace_of_tensor->get(i) += tensor.get(i, j, j) * metric.get(j, j);
-      }
-      for (size_t k = j + 1; k < dimension_of_metric; ++k) {
-        trace_of_tensor->get(i) += 2.0 * tensor.get(i, j, k) * metric.get(j, k);
-      }
-    }
-  }
-}
-
-template <typename DataType, typename Index0>
-void trace(
-    const gsl::not_null<Scalar<DataType>*> trace,
-    const Tensor<DataType, Symmetry<1, 1>, index_list<Index0, Index0>>& tensor,
-    const Tensor<DataType, Symmetry<1, 1>,
-                 index_list<change_index_up_lo<Index0>,
-                            change_index_up_lo<Index0>>>& metric) {
-  constexpr auto dimension = Index0::dim;
-  get(*trace) = get<0, 0>(tensor) * get<0, 0>(metric);
-  for (size_t i = 0; i < dimension; ++i) {
-    if (i != 0) {
-      get(*trace) += tensor.get(i, i) * metric.get(i, i);
-    }
-    for (size_t j = i + 1; j < dimension; ++j) {  // symmetry
-      get(*trace) += 2.0 * tensor.get(i, j) * metric.get(i, j);
-    }
-  }
-}
-
 #define DIM(data) BOOST_PP_TUPLE_ELEM(0, data)
 #define DTYPE(data) BOOST_PP_TUPLE_ELEM(1, data)
 #define FRAME(data) BOOST_PP_TUPLE_ELEM(2, data)
@@ -123,17 +80,7 @@ void trace(
           tensor,                                                             \
       const Tensor<DTYPE(data), Symmetry<1, 1>,                               \
                    index_list<change_index_up_lo<INDEX0(data)>,               \
-                              change_index_up_lo<INDEX0(data)>>>& metric);    \
-  template void trace_last_indices(                                           \
-      const gsl::not_null<                                                    \
-          Tensor<DTYPE(data), Symmetry<1>, index_list<INDEX0(data)>>*>        \
-          result,                                                             \
-      const Tensor<DTYPE(data), Symmetry<2, 1, 1>,                            \
-                   index_list<INDEX0(data), INDEX1(data), INDEX1(data)>>&     \
-          tensor,                                                             \
-      const Tensor<DTYPE(data), Symmetry<1, 1>,                               \
-                   index_list<change_index_up_lo<INDEX1(data)>,               \
-                              change_index_up_lo<INDEX1(data)>>>& metric);
+                              change_index_up_lo<INDEX0(data)>>>& metric);
 
 GENERATE_INSTANTIATIONS(INSTANTIATE, (1, 2, 3), (double, DataVector),
                         (Frame::Grid, Frame::Distorted, Frame::Inertial,
@@ -149,23 +96,16 @@ GENERATE_INSTANTIATIONS(INSTANTIATE, (1, 2, 3), (double, DataVector),
 #undef INDEX1
 #undef INSTANTIATE
 
-#define INSTANTIATE2(_, data)                                              \
-  template void raise_or_lower_index(                                      \
-      const gsl::not_null<                                                 \
-          Tensor<DTYPE(data), Symmetry<1>,                                 \
-                 index_list<change_index_up_lo<INDEX0(data)>>>*>           \
-          trace_of_tensor,                                                 \
-      const Tensor<DTYPE(data), Symmetry<1>, index_list<INDEX0(data)>>&    \
-          tensor,                                                          \
-      const Tensor<DTYPE(data), Symmetry<1, 1>,                            \
-                   index_list<change_index_up_lo<INDEX0(data)>,            \
-                              change_index_up_lo<INDEX0(data)>>>& metric); \
-  template void trace(                                                     \
-      const gsl::not_null<Scalar<DTYPE(data)>*> trace,                     \
-      const Tensor<DTYPE(data), Symmetry<1, 1>,                            \
-                   index_list<INDEX0(data), INDEX0(data)>>& tensor,        \
-      const Tensor<DTYPE(data), Symmetry<1, 1>,                            \
-                   index_list<change_index_up_lo<INDEX0(data)>,            \
+#define INSTANTIATE2(_, data)                                           \
+  template void raise_or_lower_index(                                   \
+      const gsl::not_null<                                              \
+          Tensor<DTYPE(data), Symmetry<1>,                              \
+                 index_list<change_index_up_lo<INDEX0(data)>>>*>        \
+          trace_of_tensor,                                              \
+      const Tensor<DTYPE(data), Symmetry<1>, index_list<INDEX0(data)>>& \
+          tensor,                                                       \
+      const Tensor<DTYPE(data), Symmetry<1, 1>,                         \
+                   index_list<change_index_up_lo<INDEX0(data)>,         \
                               change_index_up_lo<INDEX0(data)>>>& metric);
 
 GENERATE_INSTANTIATIONS(INSTANTIATE2, (1, 2, 3), (double, DataVector),
