+++ src/PointwiseFunctions/GeneralRelativity/Ricci.cpp
@@ -3,7 +3,7 @@
 
 #include "PointwiseFunctions/GeneralRelativity/Ricci.hpp"
 
-#include "DataStructures/Tensor/EagerMath/RaiseOrLowerIndex.hpp"
+#include "DataStructures/Tensor/EagerMath/Trace.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "Utilities/GenerateInstantiations.hpp"
 #include "Utilities/Gsl.hpp"
