+++ tests/Unit/DataStructures/Tensor/EagerMath/RaiseOrLowerIndex.py
@@ -6,7 +6,3 @@ import numpy as np
 
 def raise_or_lower_first_index(tensor, metric):
     return np.einsum("ij,ikl", metric, tensor)
-
-
-def trace_last_indices(tensor, metric):
-    return np.einsum("ij,kij", metric, tensor)
